Separate reducers and rename productsActions to productActions
