+++ client/src/containers/StoreView.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import StoreItem from '../components/customer/StoreItem.jsx';
 import { connect } from 'react-redux';
-import { fetchProducts } from '../actions/actions.jsx';
+import { fetchProducts } from '../actions/productActions.jsx';
 
 class StoreView extends React.Component {
   constructor(props) {
