+++ client/src/reducers/reducers.jsx
@@ -1,102 +1,10 @@
 import { combineReducers } from 'redux';
 import { reducer as formReducer } from 'redux-form';
-import {
-  ADD_PRODUCT_SUCCESS,
-  ADD_PRODUCT_FAILURE,
-  DELETE_PRODUCT_SUCCESS,
-  DELETE_PRODUCT_FAILURE,
-} from '../actions/productsActions.jsx';
-import {
-  CHANGE_VIDEO,
-  CHANGE_BROADCAST_MESSAGE,
-  SELECT_FEATURED_PRODUCT,
-} from '../actions/broadcastActions.jsx';
-import {
-  FETCH_PRODUCTS_SUCCESS,
-  FETCH_PRODUCTS_FAILURE,
-  FETCH_CUSTOMER_SUCCESS,
-  FETCH_CUSTOMER_FAILURE,
-} from '../actions/actions.jsx';
 
-function items(state = [], action) {
-  switch (action.type) {
-    case ADD_PRODUCT_SUCCESS: {
-      return [...state, action.product];
-    }
-    case ADD_PRODUCT_FAILURE: {
-      console.error(action.error);
-      return state;
-      // notify user of error?
-    }
-    case DELETE_PRODUCT_SUCCESS: {
-      return action.items;
-    }
-    case DELETE_PRODUCT_FAILURE: {
-      console.error(action.error);
-      return state;
-      // notify user of error?
-    }
-    case FETCH_PRODUCTS_SUCCESS: {
-      return action.items;
-    }
-    case FETCH_PRODUCTS_FAILURE: {
-      console.error(action.error);
-      return state;
-      // notify user of error?
-    }
-    default: {
-      return state;
-    }
-  }
-}
-function broadcastMessage(state = '', action) {
-  switch (action.type) {
-    case CHANGE_BROADCAST_MESSAGE: {
-      return action.broadcastMessage;
-    }
-    default: {
-      return state;
-    }
-  }
-}
-
-function video(state = '', action) {
-  switch (action.type) {
-    case CHANGE_VIDEO: {
-      return action.video;
-    }
-    default: {
-      return state;
-    }
-  }
-}
-
-function featuredProduct(state = false, action) {
-  switch (action.type) {
-    case SELECT_FEATURED_PRODUCT: {
-      console.log(action.product);
-      return action.product;
-    }
-    default: {
-      return state;
-    }
-  }
-}
-
-function customerInfo(state = {}, action) {
-  switch (action.type) {
-    case FETCH_CUSTOMER_SUCCESS: {
-      return action.customerInfo;
-    }
-    case FETCH_CUSTOMER_FAILURE: {
-      console.error(action.error);
-      return state;
-    }
-    default: {
-      return state;
-    }
-  }
-}
+//reducers
+import { broadcastMessage, featuredProduct, video } from './broadcastReducers.jsx';
+import { items } from './productReducers.jsx';
+import { customerInfo } from './customerReducers.jsx';
 
 const tradehouseApp = combineReducers({
   items,
